requirements.txt
# DeepEval Requirements - Complete Production Package
# Everything needed for deepeval to function fully (local + production)

# Core Dependencies (Always Required)
pandas>=1.5.0,<3.0.0
numpy>=1.21.0,<2.0.0
pydantic>=1.10.0,<3.0.0
sqlalchemy>=1.4.0,<3.0.0

# API Server (Required for API functionality)
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.20.0,<1.0.0

# LLM Provider Dependencies (For your in-house GPT integration)
openai>=1.0.0,<2.0.0
anthropic>=0.7.0,<1.0.0
google-generativeai>=0.3.0,<1.0.0

# System Monitoring
psutil>=5.9.0,<6.0.0

# Advanced ML/AI Features (Optional)
sentence-transformers>=2.2.0,<3.0.0

# Caching and Performance (Optional)
redis>=4.5.0,<5.0.0

# Development & Testing (Optional but useful)
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
black>=22.0.0,<24.0.0
isort>=5.10.0,<6.0.0